hsp-nachweis/src/main/java/de/staatsbibliothek/berlin/hsp/nachweis/domain/konfiguration/KonfigurationService.java
/*
 * MIT License
 *
 * Copyright (c) 2023 Staatsbibliothek zu Berlin - Preußischer Kulturbesitz
 *
 * Permission is hereby granted, free of charge, to any person obtaining a copy
 * of this software and associated documentation files (the "Software"), to deal
 * in the Software without restriction, including without limitation the rights
 * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
 * copies of the Software, and to permit persons to whom the Software is
 * furnished to do so, subject to the following conditions:
 *
 * The above copyright notice and this permission notice shall be included in all
 * copies or substantial portions of the Software.
 *
 * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
 * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
 * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
 * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
 * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 * SOFTWARE.
 */

package de.staatsbibliothek.berlin.hsp.nachweis.domain.konfiguration;

import java.util.Optional;
import javax.enterprise.context.ApplicationScoped;
import javax.inject.Inject;
import javax.transaction.Transactional;
import lombok.extern.slf4j.Slf4j;

/**
 * @author michael.hintersonnleitner@sbb.spk-berlin.de
 * @since 11.11.22
 */
@Slf4j
@ApplicationScoped
public class KonfigurationService implements KonfigurationBoundary {

  private final KonfigurationRepository konfigurationRepository;

  @Inject
  public KonfigurationService(KonfigurationRepository konfigurationRepository) {
    this.konfigurationRepository = konfigurationRepository;
  }

  @Override
  public Optional<String> getWert(String schluessel) {
    log.debug("getWert: schluessel={}", schluessel);
    return konfigurationRepository.getWert(schluessel);
  }

  @Override
  @Transactional
  public void setWert(String schluessel, String wert) {
    log.info("setWert: schluessel={}, wert={}", schluessel, wert);
    konfigurationRepository.setWert(schluessel, wert);
  }
}
